requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.33.0
colorama==0.4.6
constantly==15.1.0
cryptography==39.0.2
cssbeautifier==1.14.7
daphne==4.0.0
dj-database-url==2.0.0
Django==4.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django_debug_toolbar==3.8.1
django-filter==22.1
django-guardian==2.4.0
django-redis==5.3.0
django-rest-passwordreset==1.3.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djlint==1.19.16
EditorConfig==0.12.3
et-xmlfile==1.1.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsbeautifier==1.14.7
Markdown==3.4.1
msgpack==1.0.5
msgpack-python==0.5.6
numpy==1.24.3
openpyxl==3.0.10
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
pip==22.3.1
protobuf==4.23.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
redis==4.5.5
regex==2022.10.31
requests==2.31.0
rsa==4.9
service-identity==21.1.0
setuptools==65.6.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wheel==0.37.1
whitenoise==6.4.0
zope.interface==6.0
